requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
langchain==0.2.0
langchain-community==0.2.0
langchain-huggingface==0.0.3
langchain-google-genai==1.0.6
langchain_postgres==0.0.14
pdfplumber==0.11.2
torch==2.4.1
werkzeug==3.1.3